Refactor : PostErrorCode migrated to Post package - ErrorCode 도메인 패키지 아래로 이동
