+++ orury-api/src/main/java/org/fastcampus/oruryapi/domain/post/service/PostService.java
@@ -16,7 +16,7 @@ import org.fastcampus.oruryapi.domain.user.converter.dto.UserDto;
 import org.fastcampus.oruryapi.domain.user.db.repository.UserRepository;
 import org.fastcampus.oruryapi.global.constants.NumberConstants;
 import org.fastcampus.oruryapi.global.error.BusinessException;
-import org.fastcampus.oruryapi.global.error.code.PostErrorCode;
+import org.fastcampus.oruryapi.domain.post.error.PostErrorCode;
 import org.fastcampus.oruryapi.global.error.code.UserErrorCode;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
